Signal processing improvement
According `python` docs `Python signal handlers are always executed in the main Python thread`, but sometimes this is not true. This problem is too old, and on some OSs it periodically take place.
FreeBSD 10.3 amd64 send signal for last python thread, and it is not processed. FreeBSD 11.0 is not affected. Problem occurs on python 3.6 and 3.8.
Blocking signal processing in thread resolves problem independently of OS.