cli: add --use-spawn
Add a `--use-spawn` flag to the cli that will allow a user to spawn
processes instead of fork processes if they are on an OS that
defaults to forking processes such as unix based machines.

This started from a conversation on this[1] issue which showed the
problem on unix(namely macOS) systems caused by forking processes.

[1]: https://github.com/Bogdanp/dramatiq/issues/227