Add max_age as a message option
Addresses https://github.com/Bogdanp/dramatiq/issues/416

Note that the implementation is consistent with the timelimit middleware
but there is an edge case where setting a max_age of 0 will behave the
same as None, since 0 is falsey.
But this does not seem important as I cannot imagine a use-case for using
a max_age of 0. Could be fixed with explicit None checks.