Protect workers from early main process SIGHUP
Fixes #441

Note that the added testcase does not cover the issue described
in #441. The added test is a regression test only.
I am struggling to get a reliable test working that consistently
fails on master. The validation for the bugfix for #441 is manual
only.
Let me know if you would like to block merging this until an
automated test is added.