Don't reload on watchdog 'opened' events
As of watchdog >= 2.3, file open events on certain platforms are causing
an 'opened' event to be delivered.  We should not be restarting the
server on these changes, as they can be spammy and disruptive.

I was able to confirm the behaviour by running the tests within a
debian-like environment (python slim bullseye); without the fix, the
`tests/test_watch.py` would fail for the first test case (the evented observer)