Ensure only canonical are added to RabbitmqBroker.queues (#546)
When a worker is started, it'll eventually start two consumers: one for reading off the queue and another for reading off the delayed queue. When it starts consuming off the delayed queue, it calls
`RabbitmqBroker#consume('queue_name.DQ')`. In tern, 'queue_name.DQ' goes through the declare queue initialization which results in its name being added to the `queues` set. This becomes an issue as later
on in commands such as flush, assumptions are made that only the canonical queue name is present.