#651 Fix compute_backoff implementation
Ensure compute_backoff doesn't return a value smaller than `factor` (which is effectively `min_backoff`)
Do this by making `jitter` extend the backoff, rather than making it smaller.
Previously, it was this 'making it smaller' that was making it return a too small value.

Also ensure that once max_backoff is reached, the backoff is 50-100% of max_backoff, same as before.